Make use of the Rsrc Manager everywhere it should be used
now in every ressource call, the ressource manager is involved. change the Name of the Enum Variable Type in the Texture.hpp file to Format, to fit a better understanding of the usage (see Aurel's comment). Also change the call debug from std::cout to LogError to make uses of the Engine's Debug.