Engine/resource manager (#129)
* Resource Manager update with Aurel's comment

change to class definition in the cpp file, and change get lists( model and textures) to now return a referance to std :: string, and create two boolean to serve as a dirty flag to avoid calling one per frame

* Make use of the Rsrc Manager everywhere it should be used

now in every ressource call, the ressource manager is involved. change the Name of the Enum Variable Type in the Texture.hpp file to Format, to fit a better understanding of the usage (see Aurel's comment). Also change the call debug from std::cout to LogError to make uses of the Engine's Debug.

* fixed the Enum from the texture class

revert the variable name to be later fixed correctly